% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tprs.R
\name{arrangeTPRS}
\alias{arrangeTPRS}
\title{Rearrange TPRS basis}
\usage{
arrangeTPRS(tprs, intercept = FALSE)
}
\arguments{
\item{tprs}{Matrix of TPRS basis values}

\item{intercept}{Logical indicator of whether or not to remove the intercept column}
}
\description{
Rearrange the columns of the TPRS basis
}
\details{
This function takes TPRS basis values, as created by the \code{mgcv} package, and rearranges them. The last two columns are moved to the left of the matrix and the third-from last column, which corresponds to the intercept, is optionally removed.
}
\seealso{
\code{\link{computeTPRS}}
}
