% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{compute_lowCurve}
\alias{compute_lowCurve}
\title{Compute loess curves for smoothing matrix}
\usage{
compute_lowCurve(S, dgrid, newd, cl = NULL)
}
\arguments{
\item{S}{Smoothing matrix.}

\item{dgrid}{Distance matrix.}

\item{newd}{Distances to use for loess prediction.}

\item{cl}{Cluster object, or number of cluster instances to create. Defaults to no parallelization.}
}
\description{
Calculates a loess curve for the smoothing matrix entries, as a function of distance between points.
}
\seealso{
\code{\link{computeS}}
}
