% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tprs.R
\name{computeTPRS}
\alias{computeTPRS}
\title{Create TPRS basis}
\usage{
computeTPRS(coords, maxdf, rearrange = TRUE, intercept = FALSE)
}
\arguments{
\item{coords}{Data frame containing the coordinates.}

\item{maxdf}{Largest number of splines to include in TPRS basis}

\item{rearrange}{Logical indicator of whether to rearrange the columns of TPRS basis. The default from \code{mgcv} puts the linear terms at the end; re-arranging moves them to the first columns of the matrix.}

\item{intercept}{Logical indicator of whether or not to remove the intercept column}
}
\description{
Compute TPRS basis for given spatial coordinates
}
\details{
This function creates a TPRS basis using the \code{mgcv} package from the given coordinates with the option to rearrange the columns such that last two columns are moved to the left of the matrix and the third-from last column, which corresponds to the intercept, is optionally removed.
}
\examples{
x <- runif(100)
y <- runif(100)
mat <- computeTPRS(data.frame(x, y), maxdf=4)
}
\seealso{
\code{\link{arrangeTPRS}}
}
