% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{find_first_zero_cross}
\alias{find_first_zero_cross}
\title{Find first zero}
\usage{
find_first_zero_cross(x)
}
\arguments{
\item{x}{Function values, assumed to be ordered}
}
\value{
Index of first value of \code{x} that lies below 0. Decimal values will be returned using a simple interpolation of the two values straddling 0.
}
\description{
Rudimentary root finding function to calculate first cross of y-axis
}
