% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{fitLoess}
\alias{fitLoess}
\title{Fit a loess curve}
\usage{
fitLoess(y, x, newx = x, span = 0.1, ...)
}
\arguments{
\item{y}{Dependent variable values}

\item{x}{Independent variable values}

\item{newx}{Values of x to use for prediction.}

\item{span}{Passed to \code{\link[stats]{loess}}.}

\item{...}{Additional arguments passed to \code{\link[stats]{loess}}}
}
\description{
Wrapper function for fitting and predicting from \code{loess()}.
}
