% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{compute_effective_range}
\alias{compute_effective_range}
\alias{compute_effective_range_nochecks}
\title{Compute effective range}
\usage{
compute_effective_range(X, df = 3, nsamp = min(100, nrow(X)),
  newd = seq(0, 100, 1), scale_factor = 1, cl = NULL, verbose = TRUE)

compute_effective_range_nochecks(X, inds, newd, dgrid, scale_factor = 1,
  cl = NULL)
}
\arguments{
\item{X}{Matrix of spline values. Currently assumed to have columns \code{x}, \code{y}, \code{s1},\code{s2}, ...}

\item{df}{Degrees of freedom to include.}

\item{nsamp}{Number of observations from \code{X} from which to sample. Defaults to minimum of 100 and \code{nrow(X)}.}

\item{newd}{Distance values at which to make loess predictions.}

\item{scale_factor}{Factor by which range should be scaled. Usually physical distance corresponding to resolution of grid.}

\item{cl}{Cluster object, or number of cluster instances to create. Defaults to no parallelization.}

\item{verbose}{Control message printing.}

\item{inds}{Indices of observations to use for computation. Passed to \code{\link{computeS}}.}

\item{dgrid}{Distance matrix.}
}
\description{
Calculates the effective range for a spline basis matrix.
}
