% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{computeS}
\alias{computeS}
\title{Compute Smoothing Matrix}
\usage{
computeS(x, inds = 1:nrow(x))
}
\arguments{
\item{x}{Matrix of spline values. A data frame is coerced into a matrix.}

\item{inds}{Indices to compute smoothing matrix for.}
}
\value{
An \eqn{n}-by-\eqn{N}, where \eqn{n} is the length of \code{inds} and \eqn{N} is the number of rows in \code{x}.
}
\description{
Calculates the matrix \code{X(X'X)^(-1)X'}.
}
\details{
When \code{x} has many rows, this can be quite large.
}
