% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tprs.R
\name{computeTPRS}
\alias{computeTPRS}
\title{Create TPRS basis}
\usage{
computeTPRS(coords, maxdf, rearrange = TRUE)
}
\arguments{
\item{coords}{Data frame containing the coordinates.}

\item{maxdf}{Largest number of splines to include in TPRS basis}
}
\description{
Compute TPRS basis for given spatial coordinates
}
\examples{
x <- runif(100)
y <- runif(100)
mat <- computeTPRS(data.frame(x, y), maxdf=4)
}
\seealso{
\code{\link{arrangeTPRS}}
}
