% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{computeTPRS}
\alias{computeTPRS}
\title{Create TPRS basis}
\usage{
computeTPRS(coords, maxdf)
}
\arguments{
\item{coords}{Matrix of spatial coordinates. Data frame is coerced to matrix}

\item{maxdf}{Largest number of splines to include in TPRS basis}
}
\description{
Compute TPRS basis for given spatial coordinates
}
\seealso{
\code{\link{arrangeTPRS}}
}
