% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_range.R
\name{find_zeros_cross}
\alias{find_zeros_cross}
\title{Find distance to first zero}
\usage{
find_zeros_cross(D, S)
}
\arguments{
\item{D}{Distance matrix, or a subset of columns from a distance matrix.}

\item{S}{Smoothing matrix, or a subset of columns from a smoothing matrix.}
}
\value{
Vector of length equal to the number of columns in \code{D} and \code{S}. Each value is the smallest observed distance (from a column of \code{D}) that has a negative value in the corresponding column of \code{S}.
}
\description{
For a set of distance and smoothing matrix values, determines the smallest distance that corresponds with negative value for each column of the smoothing matrix.
}
